Updated local Android SDK and also completely reorganized the code into a better hierarchy of folders.
