+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mainRobotPrograms/autonomous/blue/BlueBall.java
@@ -1,14 +1,18 @@
-package org.firstinspires.ftc.teamcode.mainRobotPrograms;
+package org.firstinspires.ftc.teamcode.mainRobotPrograms.autonomous.blue;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
-@Autonomous(name="Red Ball", group = "Auto Group")
+import org.firstinspires.ftc.teamcode.mainRobotPrograms.autonomous.AutonomousBase;
 
-public class RedBall extends _AutonomousBase
+@Autonomous(name="Blue Ball", group = "Auto Group")
+
+public class BlueBall extends AutonomousBase
 {
     //Called after runOpMode() has finished initializing by BaseFunctions.
     protected void driverStationSaysGO() throws InterruptedException
     {
+        sleep(15000);
+
         //Drive to the cap ball.
         outputNewLineToDrivers ("Driving to shooting position.");
         driveForDistance (0.3, 2000);
@@ -34,7 +38,7 @@ public class RedBall extends _AutonomousBase
         outputNewLineToDrivers ("Drive to the ramp, stopping upon bttom color sensor reaches the blue region on the ramp.");
         startToDriveAt (0.25);
         long startDriveTime = System.currentTimeMillis (); //Max time at 6 seconds.
-        while (bottomColorSensor.red () <= 2.5 && (System.currentTimeMillis () - startDriveTime) < 6000)
+        while (bottomColorSensor.blue () <= 2.5 && (System.currentTimeMillis () - startDriveTime) < 6000)
             adjustMotorPowersBasedOnGyroSensor ();
         stopDriving ();
     }
